Refactor setup.py to generate module file from template and streamline source file handling; update .gitignore and MANIFEST.in for new structure
